requirements.txt
# Core Dependencies
pandas==2.0.3
numpy==1.24.3
streamlit==1.25.0
plotly==5.15.0

# Database
sqlalchemy==2.0.19

# Machine Learning
scikit-learn==1.3.0
xgboost==1.7.6
statsmodels==0.14.0

# Scientific Computing
scipy==1.11.1

# Data Generation
faker==19.3.1

# Utilities
python-dateutil==2.8.2
pyyaml==6.0.1

# Testing
pytest==7.4.0

# Optional: For advanced features
joblib==1.3.2
matplotlib==3.7.2
seaborn==0.12.2